+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend_user_service_functions.go
@@ -60,7 +60,7 @@ const (
 	userServiceContainerName = "user-service-container"
 
 	shouldMountVolumesAsReadOnly = false
-
+	shouldAddTimestampsToLogs = false
 	// Our user services don't need service accounts
 	userServiceServiceAccountName = ""
 )
@@ -359,8 +359,29 @@ func (backend *KubernetesKurtosisBackend) GetUserServiceLogs(
 	filters *service.ServiceFilters,
 	shouldFollowLogs bool,
 ) (successfulUserServiceLogs map[service.ServiceGUID]io.ReadCloser, erroredUserServiceGuids map[service.ServiceGUID]error, resultError error) {
-	//TODO implement me
-	panic("implement me")
+	serviceObjectsAndResources, err := backend.getMatchingUserServiceObjectsAndKubernetesResources(ctx, enclaveId, filters)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "Expected to be able to get user services and Kubernetes resources, instead a non nil error was returned")
+	}
+	userServiceLogs := map[service.ServiceGUID]io.ReadCloser{}
+	erredServiceLogs := map[service.ServiceGUID]error{}
+	for _, serviceObjectAndResource := range serviceObjectsAndResources {
+		serviceGuid := serviceObjectAndResource.service.GetRegistration().GetGUID()
+		servicePod := serviceObjectAndResource.kubernetesResources.pod
+		if servicePod == nil {
+			erredServiceLogs[serviceGuid] = stacktrace.NewError("Expected to find a pod for Kurtosis service with GUID '%v', instead no pod was found", serviceGuid)
+			continue
+		}
+		serviceNamespaceName := serviceObjectAndResource.kubernetesResources.service.GetNamespace()
+		// Get logs
+		logReadCloser, err := backend.kubernetesManager.GetContainerLogs(ctx, serviceNamespaceName, servicePod.Name, userServiceContainerName, shouldFollowLogs, shouldAddTimestampsToLogs)
+		if err != nil {
+			erredServiceLogs[serviceGuid] = stacktrace.Propagate(err, "Expected to be able to call Kubernetes to get logs for service with GUID '%v', instead a non-nil error was returned", serviceGuid)
+			continue
+		}
+		userServiceLogs[serviceGuid] = logReadCloser
+	}
+	return userServiceLogs, erredServiceLogs, nil
 }
 
 func (backend *KubernetesKurtosisBackend) PauseService(